+++ Packages/com.unity.ide.vscode/Editor/ProjectGeneration/FileIO.cs
@@ -1,4 +1,6 @@
+using System;
 using System.IO;
+using System.Security;
 using System.Text;
 
 namespace VSCodeEditor
@@ -11,6 +13,7 @@ namespace VSCodeEditor
         void WriteAllText(string fileName, string content);
 
         void CreateDirectory(string pathName);
+        string EscapedRelativePathFor(string file, string projectDirectory);
     }
 
     class FileIOProvider : IFileIO
@@ -33,6 +36,25 @@ namespace VSCodeEditor
         public void CreateDirectory(string pathName)
         {
             Directory.CreateDirectory(pathName);
+        }
+
+        public string EscapedRelativePathFor(string file, string projectDirectory)
+        {
+            var projectDir = Path.GetFullPath(projectDirectory);
+
+            // We have to normalize the path, because the PackageManagerRemapper assumes
+            // dir seperators will be os specific.
+            var absolutePath = Path.GetFullPath(file.NormalizePath());
+            var path = SkipPathPrefix(absolutePath, projectDir);
+
+            return SecurityElement.Escape(path);
+        }
+
+        private static string SkipPathPrefix(string path, string prefix)
+        {
+            return path.StartsWith($@"{prefix}{Path.DirectorySeparatorChar}", StringComparison.Ordinal)
+                ? path.Substring(prefix.Length + 1)
+                : path;
         }
     }
 }