+++ Packages/com.unity.ide.vscode/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Code Editor Package for Visual Studio Code
 
+## [1.2.5] - 2022-02-07
+
+- Introduce OnGeneratedCSProjectFiles, OnGeneratedCSProject and OnGeneratedSlnSolution callbacks.
+- Always use forward slash in source paths
+- Analyzers use absolute paths
+- Ruleset files for roslyn analyzers
+- Extra snap search paths on Ubuntu
+- Specific c# language version for specific unity versions
+- No longer hide .gitignore in VSCode file explorer
+
+
 ## [1.2.3] - 2020-10-23
 
 Remove workaround for VSCode omnisharp (as of https://github.com/OmniSharp/omnisharp-vscode/issues/4113 we no longer need to disable the referenceoutputassemblies).